feat(ai): update client SDK and enhance AI settings management
- Upgraded '@follow-app/client-sdk' to version 0.3.40, reflecting the latest changes in AI model management.
- Introduced new sections for managing AI shortcuts, token usage, and MCP services within the settings interface.
- Added components for personalizing prompts and managing panel styles, improving user interaction with AI features.
- Enhanced the overall layout and organization of AI settings for better usability.

These changes collectively improve the user experience by providing a more intuitive interface for managing AI settings and enhancing the overall functionality of the AI chat feature.
