+++ apps/desktop/layer/renderer/src/modules/settings/tabs/ai/index.ts
@@ -1,0 +1,5 @@
+export { MCPServicesSection } from "./mcp"
+export { PanelStyleSection } from "./PanelStyleSection"
+export { PersonalizePromptSection } from "./PersonalizePromptSection"
+export { AIShortcutsSection } from "./shortcuts"
+export { TokenUsageSection } from "./TokenUsageSection"
