+++ apps/desktop/layer/renderer/src/modules/settings/tabs/ai/mcp/index.ts
@@ -1,0 +1,3 @@
+export { MCPServiceItem } from "./MCPServiceItem"
+export { MCPServiceModalContent } from "./MCPServiceModalContent"
+export { MCPServicesSection } from "./MCPServicesSection"
