+++ apps/desktop/layer/renderer/src/modules/settings/tabs/ai/shortcuts/index.ts
@@ -1,0 +1,3 @@
+export { AIShortcutsSection } from "./AIShortcutsSection"
+export { ShortcutItem } from "./ShortcutItem"
+export { ShortcutModalContent } from "./ShortcutModalContent"
