+++ pnpm-workspace.yaml
@@ -9,7 +9,7 @@ packages:
 
 catalog:
   typescript: 5.8.3
-  "@follow-app/client-sdk": 0.3.39
+  "@follow-app/client-sdk": 0.3.40
   tailwindcss-uikit-colors: 1.0.0
 
 onlyBuiltDependencies:
